+++ src/main/scala/io/rml/framework/Main.scala
@@ -35,7 +35,11 @@ import org.apache.flink.api.java.utils.ParameterTool
 import org.apache.flink.api.scala._
 import org.apache.flink.core.fs.FileSystem.WriteMode
 import org.apache.flink.streaming.api.CheckpointingMode
-import org.apache.flink.streaming.api.scala.{DataStream, StreamExecutionEnvironment}
+import org.apache.flink.streaming.api.functions.ProcessFunction
+import org.apache.flink.streaming.api.scala.{DataStream, OutputTag, StreamExecutionEnvironment}
+import org.apache.flink.util.Collector
+
+import scala.collection.mutable
 //import org.apache.flink.streaming.connectors.kafka.{FlinkKafkaProducer09}
 
 import scala.collection.immutable
@@ -78,6 +82,9 @@ object Main extends Logging {
 
     val partitionFormat: PartitionerFormat = PartitionerFormat.fromString(partitionFormatString)
 
+    var jobName = if (parameters.has("job-name")) parameters.get("job-name")
+    else EMPTY_VALUE
+
     implicit val postProcessor:PostProcessor =
       parameters.get("post-process") match {
         case "bulk" => new BulkPostProcessor
@@ -86,7 +93,7 @@ object Main extends Logging {
       }
 
 
-    // TODO: do logging correctly
+    logInfo("Job name: " + jobName)
     logInfo("Mapping path: " + mappingPath)
     logInfo("Output path: " + outputPath)
     logInfo("Output socket: " + outputSocket)
@@ -104,13 +111,9 @@ object Main extends Logging {
     implicit val env = ExecutionEnvironment.getExecutionEnvironment
     implicit val senv = StreamExecutionEnvironment.getExecutionEnvironment
 
-
-
     senv.enableCheckpointing(5000, CheckpointingMode.AT_LEAST_ONCE);  // This is what Kafka supports ATM, see https://ci.apache.org/projects/flink/flink-docs-release-1.8/dev/connectors/guarantees.html
 
-
-    // check if the mapping contains standard dataset mappings
-    if (formattedMapping.standardTripleMaps.nonEmpty) {
+    if (formattedMapping.containsDatasetTriplesMaps() && !formattedMapping.containsStreamTriplesMaps()) {
 
       logInfo("Dataset Job Found.")
 
@@ -122,15 +125,21 @@ object Main extends Logging {
         .name("Write to output")
 
       // execute data set job
-      env.execute("DATASET JOB")
+      env.execute(jobName + " (DATASET JOB)")
 
       // check if the mapping contains streamed mappings
-    } else if (formattedMapping.streamTripleMaps.nonEmpty) {
+    } else if (formattedMapping.containsStreamTriplesMaps()) {
 
-      logInfo("Datastream Job found.")
-
-      // create a flink stream from the formatted mapping
-      val stream = createStreamFromFormattedMapping(formattedMapping)
+      val stream = if (formattedMapping.containsDatasetTriplesMaps()) {
+        logInfo("Mixed dataset and datastream job found.")
+        // At this moment, we only support the case that there is a "streaming" triples map that has a "static" parent triples map.
+        // Only the subject(s) of the parent triples map will be generated
+        createMixedPipelineFromFormattedMapping(formattedMapping)
+      } else {
+        // create a flink stream from the formatted mapping
+        logInfo("Datastream Job found.")
+        createStreamFromFormattedMapping(formattedMapping)
+      }
 
       // write to a socket if the parameter is given
       if (outputSocket != EMPTY_VALUE) stream.writeToSocket("localhost", outputSocket.toInt, new SimpleStringSchema())
@@ -145,10 +154,12 @@ object Main extends Logging {
         fact.applySink[String](kafkaBrokers,kafkaTopic, kafkaPartitionerProperties, new SimpleStringSchema(), stream)
       }
       // write to a file if the parameter is given
-      else if (!outputPath.contains(EMPTY_VALUE)) stream.writeAsText(outputPath, WriteMode.OVERWRITE)
+      else if (!outputPath.contains(EMPTY_VALUE)) {
+        stream.writeAsText(outputPath, WriteMode.OVERWRITE)
+      }
 
       // execute stream job
-      senv.execute("DATASTREAM JOB")
+      senv.execute(jobName + " (DATASTREAM JOB)")
 
     }
 
@@ -189,30 +200,30 @@ object Main extends Logging {
                                        postProcessor: PostProcessor): DataStream[String] = {
 
     // to create a Flink Data Stream there must be triple maps that contain streamed logical sources
-    require(formattedMapping.streamTripleMaps.nonEmpty)
-    val triplesMaps = formattedMapping.streamTripleMaps
+    require(formattedMapping.containsStreamTriplesMaps())
+    val triplesMaps = formattedMapping.standardStreamTriplesMaps ++ formattedMapping.joinedSteamTriplesMaps
 
     // group triple maps by logical sources based on if the postprocessor is supposed to emit at most one response
     val grouped =
       postProcessor match {
         // group by logicalSource (source + reference formulation)
-        case _:AtMostOneProcessor => triplesMaps.groupBy(tripleMap => tripleMap.logicalSource.semanticIdentifier)
+        case _:AtMostOneProcessor => triplesMaps.groupBy(triplesMap => triplesMap.logicalSource.semanticIdentifier)
 
         // group on object instance of logicalSource, i.e., don't group
-        case _:PostProcessor => triplesMaps.groupBy(tripleMap => tripleMap.logicalSource)
+        case _:PostProcessor => triplesMaps.groupBy(triplesMap => triplesMap.logicalSource)
       }
 
     // create a map with as key a Source and as value an Engine with loaded statements
     // the loaded statements are the mappings to execute
     val sourceEngineMap = grouped.map(entry => {
       var logicalSource = entry._2.head.logicalSource
-      val tripleMaps = entry._2
-      val iterators = tripleMaps.flatMap(tm => tm.logicalSource.iterators).distinct
+      val triplesMaps = entry._2
+      val iterators = triplesMaps.flatMap(tm => tm.logicalSource.iterators).distinct
       logicalSource = LogicalSource(logicalSource.referenceFormulation, iterators, logicalSource.source)
       // This creates a Source from a logical source maps this to an Engine with statements loaded from the triple maps
       Source(logicalSource) -> {
         logInfo(entry._2.size + " Triple Maps are found.")
-        StatementEngine.fromTripleMaps(tripleMaps, iterators.size > 1 )
+        StatementEngine.fromTriplesMaps(triplesMaps, iterators.size > 1 )
       }
     })
 
@@ -225,16 +236,150 @@ object Main extends Logging {
         source.stream // this will generate a stream of items
           // process every item by a processor with a loaded engine
           .map(new StdStreamProcessor(engine))
-          .name("Execute statements on items.")
+          .name("Execute mapping statements on items")
 
           // format every list of triples (as strings)
           .flatMap(list => if (list.nonEmpty) Some(list.reduce((a, b) => a + "\n" + b) + "\n\n") else None)
-          .name("Reduce to strings.")
+          .name("Convert triples to strings")
       })
 
     // union all streams to one final stream
     unionStreams(processedStreams)
 
+  }
+
+  def createMixedPipelineFromFormattedMapping(formattedMapping: FormattedRMLMapping)(implicit env: ExecutionEnvironment, senv: StreamExecutionEnvironment, postProcessor: PostProcessor): DataStream[String] = {
+    // we assume a streaming child triples map and a static parent triples map
+    require(formattedMapping.containsStreamTriplesMaps() && formattedMapping.containsDatasetTriplesMaps() && formattedMapping.containsParentTriplesMaps)
+
+    val tm2Stream = mutable.HashMap.empty[TriplesMap, DataStream[Iterable[Item]]]
+   (formattedMapping.standardStreamTriplesMaps ++ formattedMapping.joinedSteamTriplesMaps)
+     // group all mappings by logical source
+      .groupBy(_.logicalSource.semanticIdentifier)
+      .map(entry => {
+        val triplesMaps = entry._2
+        val firstSource = triplesMaps.head.logicalSource
+        val allIterators = triplesMaps
+          .flatMap( triplesMap => triplesMap.logicalSource.iterators)
+          .distinct
+        // create a new logical source with all iterators
+        val logicalSource = LogicalSource(firstSource.referenceFormulation, allIterators, firstSource.source)
+        (logicalSource, triplesMaps)
+      })
+      .foreach(logicalSource2triplesMaps => {
+        val triplesMaps = logicalSource2triplesMaps._2
+        val outputTags = triplesMaps.map(triplesMap => OutputTag[Iterable[Item]](triplesMap.hashCode().toHexString))
+
+        // Dispatch the stream to every triples map by
+        // creating a side output stream for every triples map (i.e. split the stream)
+        val logicalSource = logicalSource2triplesMaps._1
+        val dataStream = Source(logicalSource).asInstanceOf[io.rml.framework.flink.source.Stream].stream
+          .process(new ProcessFunction[Iterable[Item], Iterable[Item]] {
+            override def processElement(items: Iterable[Item], context: ProcessFunction[Iterable[Item], Iterable[Item]]#Context, out: Collector[Iterable[Item]]): Unit = {
+              // we don't need to write the items to 'out' because 
+              outputTags.foreach(context.output(_, items))
+            }
+          })
+        triplesMaps.foreach(triplesMap => {
+          val outputTag = OutputTag[Iterable[Item]](triplesMap.hashCode().toHexString)
+          val streamForTriplesMap = dataStream.getSideOutput(outputTag)
+          tm2Stream.put(triplesMap, streamForTriplesMap)
+        })
+      })
+
+    // map: (parent triples map identifier, name of the variable to join on, the value of it) => generated subject string
+    // this map contains the subjects from the static data source in order to perform the join.
+    val parentTriplesMap2JoinParentSource2JoinParentValue2Subject = getStaticParentSourceItems(formattedMapping)
+
+    val processedDataStreams: Iterable[DataStream[String]] = tm2Stream.map(entry => {
+      val triplesMap = entry._1
+      val stream = entry._2
+
+      if (triplesMap.containsParentTriplesMap) {
+        // the join scenario
+        // create the engine that will process the joined items
+        val joinedStreamTm = triplesMap.asInstanceOf[JoinedTriplesMap]
+        val engine = StatementEngine.fromJoinedTriplesMap(joinedStreamTm)
+
+        val parentTmId = joinedStreamTm.parentTriplesMap.identifier
+        val joinParentSource = joinedStreamTm.joinCondition.get.parent.identifier
+        stream
+          .flatMap(_.iterator)
+          .map(childItem => {
+            val childRef = childItem.refer(joinedStreamTm.joinCondition.get.child.identifier).get.head
+            // for every child ref from the streaming data, we look up the subject item from the static data (saved before in a map)
+            val parentItem = parentTriplesMap2JoinParentSource2JoinParentValue2Subject((parentTmId, joinParentSource, childRef))
+            // the actual join
+            val joinedItem = JoinedItem(childItem, parentItem)
+            joinedItem
+          }).name("Joining items from static data and streaming data")
+          .map(new JoinedStaticProcessor(engine))
+          .name("Executing mapping statements on joined items")
+          .flatMap(triples =>
+            if (triples.nonEmpty) {
+              Some(triples.head + "\n\n")
+            } else {
+              None
+            }
+          )
+      } else {
+
+        // the "normal" scenario.
+        val engine = StatementEngine.fromTriplesMaps(List(triplesMap))
+        stream
+          .map(new StdStreamProcessor(engine))
+          .name("Execute mapping statements on items")
+
+          // format every list of triples (as strings)
+          .flatMap(list =>
+            if (list.nonEmpty) {
+              Some(list.reduce((a, b) => a + "\n" + b) + "\n\n")
+            } else None
+          )
+          .name("Convert triples to strings")
+      }
+
+    })
+
+    unionStreams(processedDataStreams)
+  }
+
+  def getStaticParentSourceItems(formattedMapping: FormattedRMLMapping)(implicit env: ExecutionEnvironment, senv: StreamExecutionEnvironment, postProcessor: PostProcessor)
+  : Map[(String, String, String), Item] = {
+    // map: (parent triples map identiefier, name of the variable to join on, the value of it) => generated subject string
+    var parentTriplesMap2JoinParentSource2JoinParentValue2Subject = mutable.HashMap.empty[(String, String, String), Item]
+
+    formattedMapping.joinedSteamTriplesMaps.foreach(joinedTm => {
+      // identify the parent triples map
+      val parentTm = joinedTm.parentTriplesMap
+
+      // find the parent source of the join condition
+      val joinParentSource = joinedTm.joinCondition.get.parent.identifier
+
+      // get the subjects from the static logical source
+      val parentDataset = Source(joinedTm.parentTriplesMap.logicalSource).asInstanceOf[FileDataSet].dataset
+
+      parentDataset
+        // only keep items where the parent source of the join condition is defined
+        .filter(item => {
+          item.refer(joinParentSource).isDefined
+        })
+        // get the value of the join parent source and generate the subject
+        .map(item => {
+          // get the value of the parent source
+          val joinParentValue = item.refer(joinParentSource).get.head
+          //val subject = generated.map(tripleStr => tripleStr.substring(0, tripleStr.indexOf(' '))).head // should only be one subjefct!
+          (joinParentSource, joinParentValue, item)
+        })
+        // now put everything in the map
+        .collect()
+        .iterator.foreach(tuple => {
+          val parentTriplesMap2JoinId2JoinValue = (parentTm.identifier, tuple._1, tuple._2)
+          parentTriplesMap2JoinParentSource2JoinParentValue2Subject.put(parentTriplesMap2JoinId2JoinValue, tuple._3)
+        })
+    })
+
+    parentTriplesMap2JoinParentSource2JoinParentValue2Subject.toMap
   }
 
   /**
@@ -258,27 +403,27 @@ object Main extends Logging {
       * parent triple map with unique join conditions. Joined triple maps will contain only one join condition.
       * This makes it easier for setting up pipelines that need the joining of two sources.
       */
-    if (formattedMapping.standardTripleMaps.nonEmpty && formattedMapping.joinedTripleMaps.nonEmpty) {
+    if (formattedMapping.standardStaticTriplesMaps.nonEmpty && formattedMapping.joinedStaticTriplesMaps.nonEmpty) {
 
       // create a pipeline from the standard triple maps
-      val standardTMDataset = createStandardTripleMapPipeline(formattedMapping.standardTripleMaps)
+      val standardTMDataset = createStandardTriplesMapPipeline(formattedMapping.standardStaticTriplesMaps)
 
       // create a pipeline from the triple maps that contain parent triple maps
-      val tmWithPTMDataSet = createTMWithPTMPipeline(formattedMapping.joinedTripleMaps)
+      val tmWithPTMDataSet = createTMWithPTMPipeline(formattedMapping.joinedStaticTriplesMaps)
 
       // combine the two previous pipeline into one
       standardTMDataset.union(tmWithPTMDataSet)
 
       // check if the formatted mapping only contains triple maps
-    } else if (formattedMapping.standardTripleMaps.nonEmpty) {
+    } else if (formattedMapping.standardStaticTriplesMaps.nonEmpty) {
 
       // create a standard pipeline
-      createStandardTripleMapPipeline(formattedMapping.standardTripleMaps)
+      createStandardTriplesMapPipeline(formattedMapping.standardStaticTriplesMaps)
 
     } else { // the formatted mapping only contains joined triple maps
 
       // create a joined pipeline
-      createTMWithPTMPipeline(formattedMapping.joinedTripleMaps)
+      createTMWithPTMPipeline(formattedMapping.joinedStaticTriplesMaps)
     }
 
   }
@@ -286,28 +431,28 @@ object Main extends Logging {
   /**
     * Creates a pipeline from standard triple maps.
     *
-    * @param triplesMaps Triple maps which are standard.
+    * @param standardTriplesMaps Triple maps which are standard.
     * @param env         The execution environment needs to be given implicitly
     * @param senv        The execution environment needs to be given implicitly
     * @return
     */
-  private def createStandardTripleMapPipeline(triplesMaps: List[TripleMap])
+  private def createStandardTriplesMapPipeline(standardTriplesMaps: List[TriplesMap])
                                              (implicit env: ExecutionEnvironment,
                                               senv: StreamExecutionEnvironment,
                                               postProcessor: PostProcessor): DataSet[String] = {
 
     // group triple maps by logical sources
-    val grouped = triplesMaps.groupBy(tripleMap => tripleMap.logicalSource)
+    val grouped = standardTriplesMaps.groupBy(triplesMap => triplesMap.logicalSource)
 
     // create a map with as key a Source and as value an Engine with loaded statements
     // the loaded statements are the mappings to execute
     val sourceEngineMap = grouped.map(entry => {
       val logicalSource = entry._1
-      val tripleMaps = entry._2
+      val triplesMaps = entry._2
       // This creates a Source from a logical source maps this to an Engine with statements loaded from the triple maps
       Source(logicalSource) -> {
         logInfo(entry._2.size + " Triple Maps are found.")
-        StatementEngine.fromTripleMaps(tripleMaps)
+        StatementEngine.fromTriplesMaps(triplesMaps)
       }
     })
 
@@ -321,11 +466,11 @@ object Main extends Logging {
 
           // process every item by a processor with a loaded engine
           .map(new StdStaticProcessor(engine))
-          .name("Execute statements on items.")
+          .name("Execute mapping statements on items")
 
           // format every list of triples (as strings)
           .flatMap(list => if (list.nonEmpty) Some(list.reduce((a, b) => a + "\n" + b) + "\n\n") else None)
-          .name("Reduce to strings.")
+          .name("Convert triples to strings")
       })
 
     unionDataSets(processedDataSets.toList)
@@ -339,7 +484,7 @@ object Main extends Logging {
     * @param senv        The execution environment needs to be given implicitly
     * @return
     */
-  private def createTMWithPTMPipeline(triplesMaps: List[JoinedTripleMap])(implicit env: ExecutionEnvironment, senv: StreamExecutionEnvironment, postProcessor: PostProcessor): DataSet[String] = {
+  private def createTMWithPTMPipeline(triplesMaps: List[JoinedTriplesMap])(implicit env: ExecutionEnvironment, senv: StreamExecutionEnvironment, postProcessor: PostProcessor): DataSet[String] = {
     // TODO: Check if CoGroup is more efficient than Filter + Join
 
     val datasets = triplesMaps.map(tm => {
@@ -371,14 +516,14 @@ object Main extends Logging {
 
           // filter out all items that do not contain the parents join condition
           .filter(item => {
-          if (tm.joinCondition.isDefined) {
-            item.refer(tm.joinCondition.get.parent.toString).isDefined
-          } else true // if there are no join conditions all items can pass
+            if (tm.joinCondition.isDefined) {
+              item.refer(tm.joinCondition.get.parent.toString).isDefined
+            } else true // if there are no join conditions all items can pass
 
           // filter out all empty items
-        }).filter(item => {
-          !item.isInstanceOf[EmptyItem]
-        })
+          }).filter(item => {
+            !item.isInstanceOf[EmptyItem]
+          })
 
       // if there are join conditions defined join the child dataset and the parent dataset
       if (tm.joinCondition.isDefined) {
@@ -397,16 +542,16 @@ object Main extends Logging {
 
           // combine the joined item into a JoinedItem
           .map(items => {
-          val joined = JoinedItem(items._1, items._2)
-          joined
-        })
+            val joined = JoinedItem(items._1, items._2)
+            joined
+          })
 
           // process the JoinedItems in an engine
-          .map(new JoinedStaticProcessor(engine)).name("Execute statements.")
+          .map(new JoinedStaticProcessor(engine)).name("Execute mapping statements")
 
           // format the list of triples as strings
           .flatMap(list => if (list.nonEmpty) Some(list.reduce((a, b) => a + "\n" + b)) else None)
-          .name("Reduce to strings.")
+          .name("Convert triples to strings")
 
       } else { // if there are no join conditions a cross join will be executed
 
@@ -414,9 +559,9 @@ object Main extends Logging {
         val crossed = childDataset.cross(parentDataset)
 
         crossed.map(items => JoinedItem(items._1, items._2)) // create a JoinedItem from the crossed items
-          .map(new JoinedStaticProcessor(engine)).name("Execute statements.") // process the joined items
+          .map(new JoinedStaticProcessor(engine)).name("Execute mapping statements on joined items") // process the joined items
           .flatMap(list => if (list.nonEmpty) Some(list.reduce((a, b) => a + "\n" + b)) else None) // format the triples
-          .name("Reduce to strings.")
+          .name("Convert joined triples to strings")
       }