+++ src/main/scala/io/rml/framework/core/extractors/FunctionMapExtractor.scala
@@ -7,7 +7,7 @@ import io.rml.framework.shared.RMLException
 
 case class FunctionMapExtractor() extends ResourceExtractor[List[FunctionMap]] {
 
-  lazy val tripleMapExtractor: TripleMapExtractor = TripleMapExtractor()
+  lazy val triplesMapExtractor: TriplesMapExtractor = TriplesMapExtractor()
 
   /**
     * Extract.
@@ -38,9 +38,9 @@ case class FunctionMapExtractor() extends ResourceExtractor[List[FunctionMap]] {
     require(functionValues.head.isInstanceOf[RDFResource], "FunctionValue must be a resource.")
 
     val functionValue = functionValues.head.asInstanceOf[RDFResource]
-    val tripleMap = tripleMapExtractor.extractTripleMapProperties(functionValue)
-    require(tripleMap.isDefined)
-    FunctionMap(functionValue.uri.toString, tripleMap.get)
+    val triplesMap = triplesMapExtractor.extractTriplesMapProperties(functionValue)
+    require(triplesMap.isDefined)
+    FunctionMap(functionValue.uri.toString, triplesMap.get)
   }
 
 }