+++ src/main/scala/io/rml/framework/core/model/std/StdRMLMapping.scala
@@ -22,10 +22,10 @@
 
 package io.rml.framework.core.model.std
 
-import io.rml.framework.core.model.{RMLMapping, TripleMap}
+import io.rml.framework.core.model.{RMLMapping, TriplesMap}
 
-case class StdRMLMapping(triplesMaps: List[TripleMap], identifier: String) extends RMLMapping {
-  override def containsParentTripleMaps: Boolean = {
-    triplesMaps.exists(tm => tm.containsParentTripleMap)
+case class StdRMLMapping(triplesMaps: List[TriplesMap], identifier: String) extends RMLMapping {
+  override def containsParentTriplesMaps: Boolean = {
+    triplesMaps.exists(tm => tm.containsParentTriplesMap)
   }
 }